backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
google-generativeai
requests
python-dotenv
firebase-admin
neo4j
